requirements.txt
asgiref==3.8.1
Django==5.0.7
django-jazzmin==3.0.0
djangorestframework==3.15.2
drf-yasg==1.21.7
environs==11.0.0
inflection==0.5.1
marshmallow==3.21.3
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
whitenoise==6.7.0
